Move cache eviction to an external class.
The cache listener didn't really need a metadata helper, and definitely
didn't need to create an Accumulo connection. Therefore, moved the
cache management out to an external class. This also ensures all caches
in the metadata helper's cache manager are cleared upon eviction. As is,
it seems that only a subset were.