Add minimum threshold logic for finding field index holes
Update the logic for finding field index holes to support minimum
thresholds. An index hole is considered to exist for a field/datatype when
the count for an index row does not meet the threshold percentage-wise
for the count on the corresponding frequency row.

Required to support work for datawave/#825.