Add aggregator for frequency metadata rows
Create the aggregator FrequencyMetadataAggregator. This aggregator will
support compaction of the "f", "i", and "ri" columns in the metadata
table, collapsing the counts for dates into a single entry for each
unique row, data type, and column visbility grouping.

Update MetadataHelper and AllFieldsMetadataHelper such that methods
scanning over either the "f", "i", and/or "ri" columns are able to
handle entries that either have the original format created upon ingest,
or the aggregated format generated from the aggregator.

Required for datawave/issues/716.